requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.10.0=pypi_0
alabaster=0.7.12=pypi_0
alembic=1.4.3=pypi_0
apipkg=1.5=pypi_0
appdirs=1.4.4=pypi_0
atari-py=0.2.6=pypi_0
attrs=20.2.0=pypi_0
autopep8=1.5.4=pypi_0
babel=2.8.0=pypi_0
black=20.8b1=pypi_0
box2d-py=2.3.5=pypi_0
ca-certificates=2020.10.14=0
cachetools=4.1.1=pypi_0
certifi=2020.6.20=py38_0
chardet=3.0.4=pypi_0
click=7.1.2=pypi_0
cliff=3.4.0=pypi_0
cloudpickle=1.6.0=pypi_0
cmaes=0.7.0=pypi_0
cmd2=1.3.11=pypi_0
colorama=0.4.4=pypi_0
colorlog=4.4.0=pypi_0
coverage=5.3=pypi_0
cycler=0.10.0=pypi_0
dataclasses=0.6=pypi_0
dataproperty=0.50.0=pypi_0
decorator=4.4.2=pypi_0
docutils=0.16=pypi_0
execnet=1.7.1=pypi_0
flake8=3.8.4=pypi_0
future=0.18.2=pypi_0
google-auth=1.22.1=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
grpcio=1.33.1=pypi_0
gym=0.17.3=pypi_0
gym-minigrid=1.0.1=pypi_0
idna=2.10=pypi_0
imagesize=1.2.0=pypi_0
importlab=0.5.1=pypi_0
iniconfig=1.1.1=pypi_0
isort=5.6.4=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.17.0=pypi_0
kiwisolver=1.2.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
livereload=2.6.3=pypi_0
mako=1.1.3=pypi_0
markdown=3.3.3=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.2=pypi_0
mbstrdecoder=1.0.0=pypi_0
mccabe=0.6.1=pypi_0
msgfy=0.1.0=pypi_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.2=he6710b0_1
networkx=2.5=pypi_0
ninja=1.10.0.post2=pypi_0
numpy=1.19.2=pypi_0
oauthlib=3.1.0=pypi_0
opencv-python=4.4.0.44=pypi_0
openssl=1.1.1h=h7b6447c_0
optuna=2.2.0=pypi_0
packaging=20.4=pypi_0
pandas=1.1.3=pypi_0
pathspec=0.8.0=pypi_0
pathvalidate=2.3.0=pypi_0
pbr=5.5.1=pypi_0
pillow=8.0.1=pypi_0
pip=20.2.4=py38_0
pluggy=0.13.1=pypi_0
prettytable=0.7.2=pypi_0
protobuf=3.13.0=pypi_0
psutil=5.7.3=pypi_0
py=1.9.0=pypi_0
pyaml=20.4.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pybullet=3.0.6=pypi_0
pycodestyle=2.6.0=pypi_0
pyenchant=3.1.1=pypi_0
pyflakes=2.2.0=pypi_0
pyglet=1.5.0=pypi_0
pygments=2.7.2=pypi_0
pyparsing=2.4.7=pypi_0
pyperclip=1.8.1=pypi_0
pytablewriter=0.58.0=pypi_0
pytest=6.1.1=pypi_0
pytest-cov=2.10.1=pypi_0
pytest-env=0.6.2=pypi_0
pytest-forked=1.3.0=pypi_0
pytest-xdist=2.1.0=pypi_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=pypi_0
python-editor=1.0.4=pypi_0
pytype=2020.10.8=pypi_0
pytz=2020.1=pypi_0
pyyaml=5.3.1=pypi_0
readline=8.0=h7b6447c_0
regex=2020.10.23=pypi_0
requests=2.24.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
sb3-contrib=0.10.0a1=pypi_0
scikit-learn=0.23.2=pypi_0
scikit-optimize=0.8.1=pypi_0
scipy=1.5.3=pypi_0
seaborn=0.11.0=pypi_0
setuptools=50.3.0=py38hb0f4dca_1
sida=0.1.3=pypi_0
six=1.15.0=pypi_0
snowballstemmer=2.0.0=pypi_0
sphinx=3.2.1=pypi_0
sphinx-autobuild=2020.9.1=pypi_0
sphinx-autodoc-typehints=1.11.1=pypi_0
sphinx-rtd-theme=0.5.0=pypi_0
sphinxcontrib-applehelp=1.0.2=pypi_0
sphinxcontrib-devhelp=1.0.2=pypi_0
sphinxcontrib-htmlhelp=1.0.3=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.3=pypi_0
sphinxcontrib-serializinghtml=1.1.4=pypi_0
sphinxcontrib-spelling=7.0.0=pypi_0
sqlalchemy=1.3.20=pypi_0
sqlite=3.33.0=h62c20be_0
stable-baselines3=0.10.0a1=pypi_0
stevedore=3.2.2=pypi_0
tabledata=1.1.3=pypi_0
tcolorpy=0.0.8=pypi_0
tensorboard=2.3.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
toml=0.10.1=pypi_0
torch=1.7.0+cu101=pypi_0
torchaudio=0.7.0=pypi_0
torchvision=0.8.1+cu101=pypi_0
tornado=6.0.4=pypi_0
tqdm=4.51.0=pypi_0
typed-ast=1.4.1=pypi_0
typepy=1.1.1=pypi_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.25.11=pypi_0
wcwidth=0.2.5=pypi_0
websocket-client=0.57.0=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.35.1=py_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
